Refactor for request and response location, re-moving it outside handler
